+++ src/lib/config.ts
@@ -58,8 +58,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "2025.9.21",
-  genVersion: "2.708.2",
+  sdkVersion: "2025.9.23",
+  genVersion: "2.709.0",
   userAgent:
-    "speakeasy-sdk/typescript 2025.9.21 2.708.2 1.0.0 @wistia/wistia-api-client",
+    "speakeasy-sdk/typescript 2025.9.23 2.709.0 1.0.0 @wistia/wistia-api-client",
 } as const;